Revisions
I merged the Material Design with my changes so that it works as it
should.

Additionally, I created some additional things for the Watch class. Now
it'll be really easy to add more watches just by adding them to the
WatchManager and roll it out to the preview card as well as the creator
modal.

I put in a few new sliders where appropriate, giving them logical values