Inputs for sliders, cheatsheet and parameterlists
Created all the parameter lists, can be found in seperate html file for
you guys to copy.
Created a cheatsheet for the tags, as they can be used anywhere i made
it an persistant panel. maybe we need to revise this. Also i added
inputs and linked them to the sliders to users can enter a number and
not fiddle with the slider.

Fixed some bugs.
The only issue i have is that layer creation in the canvas is broken
now. One can not select and edit a layer anymore.